[frontend]Improvements and removed url,onfilepathchange from OverflowingItem
