+++ desktop/core/src/desktop/js/reactComponents/FileChooser/PathBrowser/Breadcrumb/Breadcrumb.scss
@@ -21,3 +21,4 @@
     min-width: 4ch;
   }
 }
+
